Make the pre-set delays longer, remove one
It turns out that we already have a gap at 3pm on Sunday since that's
when the knockouts start. Since we now also have fewer matches,
remove that delay and move the end of the league forwards a bit.
Up the other delays to the requested 10 minutes now that I've worked
out why that was causing errors previously (it was the delay in the
knockouts, which is now removed).

Change-Id: Ia9ae135fb769181b6f606fad3343a5207d776ec1